Update logic.lua
This adds logic to remove the ledge near Pewter. A lot of AP versions ago there was a ledge leading from Digletts Cave to Pewter so softlocks could be prevented. When the "Pallet Warp" got added, they reverted the ledge back to a Cut Tree but the poptracker pack never got this change.